move Scala3 to own dir
