+++ packages/quicktype-core/src/language/Scala3/index.ts
@@ -1,0 +1,4 @@
+export { Scala3TargetLanguage, scala3Options } from "./language";
+export { Scala3Renderer } from "./Scala3Renderer";
+export { CirceRenderer } from "./CirceRenderer";
+export { UpickleRenderer } from "./UpickleRenderer";
